jetls-client: Split the client into per-component modules
`jetls-client.ts` had grown to roughly 900 lines mixing the extension
entry points with the status bar, configuration handling, the server
lifecycle, and update checks, while only small vscode-free helpers
lived in separate files. This made the layout look arbitrary: one tiny
utility (`CoalescingTaskRunner`) had its own file while everything
else shared one.

This commit splits the client per component: `jetls-client.ts` keeps
only `activate`/`deactivate` (the esbuild entry name is preserved),
`server-lifecycle.ts` owns the server start/restart/shutdown flow,
`status-bar.ts` wraps the startup status bar in a `StartupStatusBar`
class, `server-config.ts` reads and compares the server configuration,
and `update-check.ts` hosts the version update checks. It also renames
`server-startup.ts` to `preflight.ts` and `CoalescingTaskRunner.ts` to
`coalescing-task-runner.ts`, following the file naming convention.

The moves are verbatim except for the status bar: the `deactivating`
guard inside the old `showServerStartupStatus` is now a `suppress()`
method on `StartupStatusBar`, called at the start of `deactivate`.
`CoalescingTaskRunner` stays in its own vscode-free module because
modules imported from `test/*.test.ts` are bundled without `vscode`
as an external, so unit-tested code cannot depend on it.
